refactor(internal): split smart_search.py into focused mixin modules
- smart_search.py (2478 lines) becomes a 63-line shell: SmartSearchTools
  composes DeepSearchMixin/SystemOverviewMixin/EntitySearchMixin, keeps
  __init__ + create_smart_search_tools so import and monkeypatch paths
  (get_global_settings) are unchanged
- New siblings: _search_config (constants, dependency-free), _search_base
  (shared attrs + registry-list helper), _search_deep, _search_overview,
  _search_entities
- Behavior-preserving: methods moved verbatim; AUTOMATION_CONFIG_TIME_BUDGET
  test patch repointed to _search_deep; docstring-harvester paths updated
- Document module-size guidance + the mixin-split pattern in AGENTS.md
