+++ AGENTS.md
@@ -661,6 +661,10 @@ This project's tool count exceeds the [10-20 tool threshold](https://ai.google.d
 
 **Context engineering**: provide minimum context; let models fetch more via `ha_get_skill_guide`. Favor statelessness and content-derived hashes for optimistic locking.
 
+### Module Size
+
+Keep modules focused. Past ~1000 lines (Pylint's `max-module-lines` default) a module usually spans multiple concerns and is worth splitting. To split a large class, prefer feature mixins over one giant class: a thin shell module keeps the public class + `__init__` (preserving import and monkeypatch paths) and composes mixins from sibling `_<area>_*.py` files, with shared constants in a dependency-free `_*_config.py` to avoid import cycles. Example: `smart_search.py` (shell) + `_search_{base,config,deep,overview,entities}.py`.
+
 ## Tool Waiting Behavior
 
 **Principle**: MCP tools should wait for operations to complete before returning, not just acknowledge API success.
